requirements.txt
asgiref==3.8.1
base58==2.1.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
Django==5.1.2
djangorestframework==3.15.2
idna==3.10
msgpack==1.1.0
multiaddr==0.0.9
netaddr==1.3.0
py-algorand-sdk==2.6.1
pycparser==2.22
pycryptodomex==3.21.0
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
varint==1.0.2
